move module into hypervisor
