+++ src/hypervisor/mod.rs
@@ -2,16 +2,18 @@ use alloc::sync::Arc;
 use alloc::vec::Vec;
 use spin::Mutex;
 
-use crate::hyp_alloc::FrameTracker;
 use crate::sync::UPSafeCell;
 use crate::mm::MemorySet;
 use crate::page_table::{PageTable, PageTableSv39};
 use crate::debug::PageDebug;
+pub use self::hyp_alloc::FrameTracker;
 
 use lazy_static::lazy_static;
 
 
-mod mm;
+pub mod device;
+pub mod hyp_alloc;
+pub mod trap;
 
 pub struct Hypervisor<P: PageTable + PageDebug> {
     pub hyper_space: Arc<UPSafeCell<MemorySet<P>>>,
