+++ src/hypervisor/trap/context.rs
