+++ src/hypervisor/trap/device.rs
