+++ src/main.rs
@@ -27,19 +27,16 @@ mod page_table;
 mod sbi;
 mod sync;
 mod timer;
-pub mod trap;
 mod guest;
 mod debug;
-mod hyp_alloc;
 mod mm;
 mod device_emu;
-mod device;
 mod hypervisor;
 
 
 
 use crate::constants::layout::PAGE_SIZE;
-use crate::device::init_dt;
+use crate::hypervisor::device::init_dt;
 use crate::guest::{GuestKernel, GUEST_KERNEL_MANAGER, run_guest_kernel};
 use crate::mm::MemorySet;
 
@@ -99,15 +96,15 @@ pub fn hentry(hart_id: usize, device_tree_blob: usize) -> ! {
         hdebug!("hart_id: {}, device tree blob: {:#x}", hart_id, device_tree_blob);
         init_dt(device_tree_blob);
         // 初始化堆及帧分配器
-        hyp_alloc::heap_init();
+        hypervisor::hyp_alloc::heap_init();
         let guest_kernel_memory = MemorySet::new_guest_kernel(&GUEST_KERNEL);
         // 初始化虚拟内存
         mm::vm_init(&guest_kernel_memory);
-        trap::init();
+        hypervisor::trap::init();
         mm::remap_test();
         mm::guest_kernel_test();
         // 开启时钟中断
-        trap::enable_timer_interrupt();
+        hypervisor::trap::enable_timer_interrupt();
         timer::set_default_next_trigger();
         // 创建用户态的 guest kernel 内存空间
         let user_guest_kernel_memory = MemorySet::create_user_guest_kernel(&guest_kernel_memory);
